Add DbComment type
Topics and comment bodies in `Comment` should use the `Topic` and `CommentText` types, rather than
just `Text`. Create a `DbComment` type that corresponds to what we get out of the database, and a
conversion function that gives us an `Error` or a valid `Comment`.

When we're viewing comments, we can do the conversion on each `DbComment` and then filter using
`rights`.